feat: add support for Windows platform

+++ src-tauri/src/lib.rs
@@ -1,5 +1,6 @@
 use std::collections::HashMap;
 use std::path::PathBuf;
+use std::sync::mpsc;
 use std::{
     fs,
     sync::{Arc, Mutex},
@@ -27,13 +28,78 @@ struct ModelInfo {
     status: String,
 }
 
+#[derive(Clone)]
 struct AppState {
+    audio_output: Arc<Mutex<AudioOutput>>,
     whisper: Arc<Mutex<Option<Whisper>>>,
     translator: Arc<Mutex<Option<Translator>>>,
 }
 
 impl AppState {
+    pub fn new(app: AppHandle) -> anyhow::Result<Self> {
+        let (audio_sender, audio_receiver) = mpsc::channel();
+        let (transcript_sender, transcript_receiver) = mpsc::channel();
+        let audio_output = AudioOutput::new(audio_sender)?;
+        let whisper = Arc::new(Mutex::new(None::<Whisper>));
+        let whisper_arc = whisper.clone();
+        let translator = Arc::new(Mutex::new(None::<Translator>));
+        let translator_arc = translator.clone();
+
+        std::thread::spawn(move || {
+            while let Ok(samples) = audio_receiver.recv() {
+                let mut whisper = whisper_arc.lock().unwrap();
+                if whisper.is_none() {
+                    continue;
+                }
+                let text = whisper.as_mut().unwrap().transcribe(samples).unwrap();
+                transcript_sender.send(text).unwrap();
+            }
+        });
+
+        std::thread::spawn(move || {
+            while let Ok(text) = transcript_receiver.recv() {
+                let mut translator = translator_arc.lock().unwrap();
+                if translator.is_none() {
+                    continue;
+                }
+                let translated_text = translator.as_mut().unwrap().translate(&text).unwrap();
+                if text == " [BLANK_AUDIO]" {
+                    app.emit(
+                        "event",
+                        Event {
+                            original_text: "BLANK_AUDIO".to_string(),
+                            translated_text: "空白".to_string(),
+                        },
+                    )
+                    .unwrap();
+                    continue;
+                }
+                log::debug!("original_text: {}", text);
+                log::debug!("translated_text: {}", translated_text);
+                app.emit(
+                    "event",
+                    Event {
+                        original_text: text,
+                        translated_text,
+                    },
+                )
+                .unwrap();
+            }
+        });
+
+        Ok(Self {
+            audio_output: Arc::new(Mutex::new(audio_output)),
+            whisper,
+            translator,
+        })
+    }
+
+    fn is_ready(&self) -> bool {
+        self.whisper.lock().unwrap().is_some() && self.translator.lock().unwrap().is_some()
+    }
+
     fn set_model(&self, app: &AppHandle, file_name: &str) -> Result<(), String> {
+        log::info!("create model: {}", file_name);
         match file_name {
             "ggml-base-q5_1.bin" => {
                 self.whisper
@@ -50,10 +116,6 @@ impl AppState {
             _ => unreachable!(),
         }
         Ok(())
-    }
-
-    fn is_ready(&self) -> bool {
-        self.whisper.lock().unwrap().is_some() && self.translator.lock().unwrap().is_some()
     }
 
     fn create_whisper(app: &AppHandle, file_name: &str) -> Result<Whisper, String> {
@@ -101,131 +163,71 @@ pub struct DownloadProgress {
 #[tauri::command]
 async fn start_recording(
     app: AppHandle,
-    output: tauri::State<'_, AudioOutput>,
     state: tauri::State<'_, AppState>,
 ) -> Result<bool, String> {
     log::info!("start_recording");
     if !state.is_ready() {
-        open_settings(app)?;
+        open_settings(app).await?;
         return Ok(false);
     }
-    if !output.start_recording() {
-        return Ok(false);
-    }
+    app.emit(
+        "event",
+        Event {
+            original_text: "wait for audio".to_string(),
+            translated_text: "等待音频".to_string(),
+        },
+    )
+    .unwrap();
 
-    let mut rx = output.sender.subscribe();
-
-    let output_clone = output.inner().clone();
-    let whisper = state.whisper.lock().unwrap().as_ref().unwrap().clone();
-    let translator = state.translator.lock().unwrap().as_ref().unwrap().clone();
-    tauri::async_runtime::spawn(async move {
-        let (tx, mut rx1) = tokio::sync::mpsc::channel(1);
-        let whisper_clone = whisper.clone();
-        let collect_task = tokio::spawn(async move {
-            while let Ok(sample_buf) = rx.recv().await {
-                if output_clone.is_stopped() || sample_buf.is_none() {
-                    log::info!(
-                        "stop by user {} | sample is none {}",
-                        output_clone.is_stopped(),
-                        sample_buf.is_none()
-                    );
-                    tx.send(false).await.unwrap();
-                    break;
-                }
-                let audio_buf_list = sample_buf.unwrap().audio_buf_list::<2>().unwrap();
-                let buffer_list = audio_buf_list.list();
-                let samples = unsafe {
-                    let buffer = buffer_list.buffers[0];
-                    std::slice::from_raw_parts(
-                        buffer.data as *const f32,
-                        buffer.data_bytes_size as usize / std::mem::size_of::<f32>(),
-                    )
-                };
-                whisper_clone.add_new_samples(samples, 48000, 1);
-                if whisper_clone.can_transcribe() {
-                    tx.send(true).await.unwrap()
-                }
-            }
-            log::info!("# Stop whisper");
-        });
-        let whisper_clone = whisper.clone();
-        let (tx1, mut rx2) = tokio::sync::mpsc::channel(1);
-        let transcribe_task = tokio::spawn(async move {
-            while (rx1.recv().await).is_some() {
-                if let Some(text) = whisper_clone.transcribe() {
-                    tx1.send(Some(text)).await.unwrap();
-                }
-            }
-            tx1.send(None).await.unwrap();
-        });
-        let translator_clone = translator.clone();
-        let translate_task = tokio::spawn(async move {
-            while let Some(Some(text)) = rx2.recv().await {
-                if text == " [BLANK_AUDIO]" {
-                    app.emit(
-                        "event",
-                        Event {
-                            original_text: "BLANK_AUDIO".to_string(),
-                            translated_text: "空白".to_string(),
-                        },
-                    )
-                    .unwrap();
-                    continue;
-                }
-                let translated_text = translator_clone.translate(&text).unwrap();
-                log::debug!("original_text: {}", text);
-                log::debug!("translated_text: {}", translated_text);
-                app.emit(
-                    "event",
-                    Event {
-                        original_text: text,
-                        translated_text,
-                    },
-                )
-                .unwrap();
-            }
-            log::info!("# Stop translate");
-            app.emit(
-                "event",
-                Event {
-                    original_text: "".to_string(),
-                    translated_text: "".to_string(),
-                },
-            )
-            .unwrap();
-        });
-        let _ = tokio::join!(collect_task, transcribe_task, translate_task);
-    });
-
+    state
+        .audio_output
+        .lock()
+        .unwrap()
+        .start_recording()
+        .map_err(|e| e.to_string())?;
     Ok(true)
 }
 
 #[tauri::command]
-fn stop_recording(output: tauri::State<'_, AudioOutput>) -> Result<(), String> {
+fn stop_recording(app: AppHandle, state: tauri::State<'_, AppState>) -> Result<(), String> {
     log::info!("stop_recording");
-    output.stop_recording();
+    state.audio_output.lock().unwrap().stop_recording();
+    app.emit(
+        "event",
+        Event {
+            original_text: "已暂停".to_string(),
+            translated_text: "".to_string(),
+        },
+    )
+    .unwrap();
     Ok(())
 }
 
 #[tauri::command]
-fn open_settings(app: AppHandle) -> Result<(), String> {
+async fn open_settings(app: AppHandle) -> Result<(), String> {
     // Check if settings window already exists and focus it
     if let Some(settings_window) = app.get_webview_window("settings") {
         settings_window.set_focus().map_err(|e| e.to_string())?;
         Ok(())
     } else {
-        let settings = WebviewWindowBuilder::new(
+        let mut builder = WebviewWindowBuilder::new(
             &app,
             "settings",
             tauri::WebviewUrl::App("/#/settings".into()),
         )
-        .hidden_title(true)
-        .title_bar_style(tauri::TitleBarStyle::Overlay)
         .inner_size(400.0, 300.0)
         .resizable(false)
-        .center()
-        .build();
+        .title("Model Download")
+        .center();
 
+        #[cfg(target_os = "macos")]
+        {
+            builder = builder
+                .title_bar_style(tauri::TitleBarStyle::Overlay)
+                .hidden_title(true);
+        }
+
+        let settings = builder.build();
         match settings {
             Ok(_) => Ok(()),
             Err(e) => Err(e.to_string()),
@@ -242,42 +244,51 @@ async fn download_model(
 ) -> Result<(), String> {
     let model_dir = model_dir(&app)?;
     let file_path = model_dir.join(&file_name);
+    log::info!("download model save to: {}", file_path.to_str().unwrap());
+    let mut total_size = 0;
+    let mut downloaded = 0;
 
-    // Download the file
-    let response = reqwest::get(&url).await.map_err(|e| e.to_string())?;
+    if file_path.exists() {
+        log::info!("model already exists: {}", file_path.to_str().unwrap());
+        // TODO: check file md5
+    } else {
+        // Download the file
+        let response = reqwest::get(&url).await.map_err(|e| e.to_string())?;
 
-    let total_size = response
-        .content_length()
-        .ok_or_else(|| "Failed to get content length".to_string())?;
-    let mut file = fs::File::create(&file_path).map_err(|e| e.to_string())?;
-    let mut downloaded: u64 = 0;
-    let mut stream = response.bytes_stream();
+        total_size = response
+            .content_length()
+            .ok_or_else(|| "Failed to get content length".to_string())?;
+        let mut file = fs::File::create(&file_path).map_err(|e| e.to_string())?;
+        let mut stream = response.bytes_stream();
 
-    use futures_util::StreamExt;
-    let mut last_update = std::time::Instant::now();
-    while let Some(item) = stream.next().await {
-        let chunk = item.map_err(|e| e.to_string())?;
-        std::io::copy(&mut chunk.as_ref(), &mut file).map_err(|e| e.to_string())?;
-        downloaded += chunk.len() as u64;
+        use futures_util::StreamExt;
+        let mut last_update = std::time::Instant::now();
+        while let Some(item) = stream.next().await {
+            let chunk = item.map_err(|e| e.to_string())?;
+            std::io::copy(&mut chunk.as_ref(), &mut file).map_err(|e| e.to_string())?;
+            downloaded += chunk.len() as u64;
 
-        // Only update progress every 200ms
-        let now = std::time::Instant::now();
-        if now.duration_since(last_update).as_millis() >= 200 {
-            let progress = (downloaded as f32 / total_size as f32) * 100.0;
-            app.emit(
-                "download-progress",
-                DownloadProgress {
-                    file_name: file_name.clone(),
-                    progress,
-                    total_size,
-                    downloaded,
-                },
-            )
-            .map_err(|e| e.to_string())?;
-            last_update = now;
+            // Only update progress every 200ms
+            let now = std::time::Instant::now();
+            if now.duration_since(last_update).as_millis() >= 200 {
+                let progress = (downloaded as f32 / total_size as f32) * 100.0;
+                log::info!("download progress: {} %", progress);
+                app.emit(
+                    "download-progress",
+                    DownloadProgress {
+                        file_name: file_name.clone(),
+                        progress,
+                        total_size,
+                        downloaded,
+                    },
+                )
+                .map_err(|e| e.to_string())?;
+                last_update = now;
+            }
         }
     }
     state.set_model(&app, &file_name)?;
+    log::debug!("download model: {} completed", file_name);
     app.emit(
         "download-progress",
         DownloadProgress {
@@ -307,14 +318,21 @@ fn model_dir(app: &AppHandle) -> Result<PathBuf, String> {
     Ok(model_dir)
 }
 
+#[tauri::command]
+async fn show_main_window(window: tauri::Window) {
+    window.get_window("main").unwrap().show().unwrap();
+}
+
 #[cfg_attr(mobile, tauri::mobile_entry_point)]
 pub fn run() {
     tauri::Builder::default()
+        .plugin(tauri_plugin_log::Builder::new().build())
         .plugin(tauri_plugin_store::Builder::new().build())
         .plugin(
             tauri_plugin_log::Builder::new()
                 .level(log::LevelFilter::Debug)
                 .targets(vec![
+                    tauri_plugin_log::Target::new(tauri_plugin_log::TargetKind::Webview),
                     tauri_plugin_log::Target::new(tauri_plugin_log::TargetKind::Stdout),
                     tauri_plugin_log::Target::new(tauri_plugin_log::TargetKind::LogDir {
                         file_name: Some("Peeches".to_string()),
@@ -344,8 +362,8 @@ pub fn run() {
             {
                 // Hide dock icon
                 app.set_activation_policy(tauri::ActivationPolicy::Accessory);
-                window.set_shadow(false)?;
             }
+            window.set_shadow(false)?;
 
             if let Some(monitor) = window.primary_monitor().unwrap() {
                 window.set_min_size(Some(tauri::Size::Physical(tauri::PhysicalSize {
@@ -378,6 +396,8 @@ pub fn run() {
                     .unwrap();
             }
 
+            let app_state = AppState::new(app.handle().clone())?;
+
             let model_dir = model_dir(app.handle())?;
 
             // Get model store state
@@ -388,65 +408,42 @@ pub fn run() {
             let mut models: HashMap<String, ModelInfo> =
                 serde_json::from_value(models).unwrap_or_default();
 
-            let whisper_model = {
-                if let Some(info) = models.get("ggml-base-q5_1.bin") {
-                    if info.status == "completed" {
-                        let model_path = model_dir.join(&info.file_name);
-                        if model_path.exists() {
-                            let whisper = Whisper::new(model_path.to_str().unwrap());
-                            Some(whisper)
-                        } else {
-                            models.remove("ggml-base-q5_1.bin");
-                            store.set("models", serde_json::to_value(&models).unwrap());
-                            None
-                        }
-                    } else {
-                        None
-                    }
+            if let Some(info) = models.get("ggml-base-q5_1.bin") {
+                let model_path = model_dir.join(&info.file_name);
+                if info.status == "completed" && model_path.exists() {
+                    let whisper = Whisper::new(model_path.to_str().unwrap());
+                    app_state.whisper.lock().unwrap().replace(whisper);
                 } else {
-                    None
+                    models.remove("ggml-base-q5_1.bin");
+                    store.set("models", serde_json::to_value(&models).unwrap());
                 }
             };
 
-            let translator = {
-                if let Some(info) = models.get("opus-mt-en-zh.bin") {
-                    if info.status == "completed" {
-                        let model_path = model_dir.join(&info.file_name);
-                        if model_path.exists() {
-                            let (en_token, zh_token) = get_token_path(app.handle());
-                            let translator = Translator::new(
-                                model_path.to_str().unwrap(),
-                                en_token.to_str().unwrap(),
-                                zh_token.to_str().unwrap(),
-                            )?;
-                            Some(translator)
-                        } else {
-                            models.remove("opus-mt-en-zh.bin");
-                            store.set("models", serde_json::to_value(&models).unwrap());
-                            None
-                        }
-                    } else {
-                        None
-                    }
+            if let Some(info) = models.get("opus-mt-en-zh.bin") {
+                let model_path = model_dir.join(&info.file_name);
+                if info.status == "completed" && model_path.exists() {
+                    let (en_token, zh_token) = get_token_path(app.handle());
+                    let translator = Translator::new(
+                        model_path.to_str().unwrap(),
+                        en_token.to_str().unwrap(),
+                        zh_token.to_str().unwrap(),
+                    )?;
+                    app_state.translator.lock().unwrap().replace(translator);
                 } else {
-                    None
+                    models.remove("opus-mt-en-zh.bin");
+                    store.set("models", serde_json::to_value(&models).unwrap());
                 }
             };
-
-            app.manage(AppState {
-                whisper: Arc::new(Mutex::new(whisper_model)),
-                translator: Arc::new(Mutex::new(translator)),
-            });
-
+            app.manage(app_state);
             Ok(())
         })
-        .manage(AudioOutput::new())
         .invoke_handler(tauri::generate_handler![
             close_app,
             start_recording,
             stop_recording,
             open_settings,
-            download_model
+            download_model,
+            show_main_window
         ])
         .run(tauri::generate_context!())
         .expect("error while running tauri application");
